+++ app/src/main/java/apps/czeidler/economylogboook/data/DataManager.java
@@ -85,10 +85,10 @@ public class DataManager extends SQLiteOpenHelper {
 
     private EconEntry buildEntry(Cursor cursor) {
         return new EconEntry(cursor.getLong(4),
-                cursor.getFloat(2), // fuel
-                cursor.getFloat(3), // fuel mult
-                cursor.getFloat(0), // dist
-                cursor.getFloat(1)); //dist mult
+                cursor.getDouble(2), // fuel
+                cursor.getDouble(3), // fuel mult
+                cursor.getDouble(0), // dist
+                cursor.getDouble(1)); //dist mult
     }
 
     public void deleteEntries() {
